Added OpenRefine project.
Currently, due to the fact that OpenRefine does not include single cell
operations in its JSON, the data set rebuild from the exported data does
not match the data set at time of export. Therefore, we include the full
OpenRefine project so that the data set can be rebuild to the expected
state.